fix(utils): code cleanup
